Renamed UserRepository to UserDAO, renamed package as well
