+++ src/main/java/com/atspickem/backend/BackendApplication.java
@@ -1,12 +1,12 @@
 package com.atspickem.backend;
 
-import com.atspickem.backend.repository.UserRepository;
+import com.atspickem.backend.dao.UserDAO;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @SpringBootApplication
-@EnableJpaRepositories(basePackageClasses = UserRepository.class)
+@EnableJpaRepositories(basePackageClasses = UserDAO.class)
 public class BackendApplication {
 
 	public static void main(String[] args) {
