+++ src/main/java/com/atspickem/backend/services/MyUserDetailsService.java
@@ -1,7 +1,7 @@
 package com.atspickem.backend.services;
 
 
-import com.atspickem.backend.repository.UserRepository;
+import com.atspickem.backend.dao.UserDAO;
 import com.atspickem.backend.models.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -15,11 +15,11 @@ import java.util.Optional;
 public class MyUserDetailsService implements UserDetailsService {
 
     @Autowired
-    UserRepository userRepository;
+    UserDAO userDAO;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Optional<User> user = Optional.ofNullable(userRepository.findByUsername(username));
+        Optional<User> user = Optional.ofNullable(userDAO.findByUsername(username));
 
         user.orElseThrow(() -> new UsernameNotFoundException("Not found " + username));
